feat: Implement home layout and components including HomeHeader, FollowingPostsTimeline, and GlobalPostsTimeline
