+++ src/app/[locale]/(home)/_components/FollowingPostsTimeline/index.ts
