+++ src/app/[locale]/(home)/_components/GlobalPostsTimeline/GlobalPostsTimeline.tsx
