+++ src/app/[locale]/(home)/_components/GlobalPostsTimeline/index.ts
