+++ src/app/[locale]/(home)/_components/HomeHeader/config.ts
