+++ src/app/[locale]/(home)/page.tsx
@@ -7,15 +7,13 @@ import { HomeTab, useHomeStore } from '@/stores/home';
 
 import FollowingPostsTimeline from './_components/FollowingPostsTimeline';
 import GlobalPostsTimeline from './_components/GlobalPostsTimeline';
-import HomeTabBar from './_components/HomeTabBar';
 
 export default function Home() {
   const { user } = useAppSession();
   const currentTab = useHomeStore((state) => state.currentTab);
 
   return (
-    <Box className='min-h-screen gap-0'>
-      <HomeTabBar />
+    <>
       {user && (
         <Box className='hidden gap-0 border-b border-border-1 sm:flex'>
           <CreatePostForm />
@@ -23,6 +21,6 @@ export default function Home() {
       )}
       {currentTab === HomeTab.GLOBAL && <GlobalPostsTimeline />}
       {currentTab === HomeTab.FOLLOWING && <FollowingPostsTimeline />}
-    </Box>
+    </>
   );
 }
