+++ src/app/[locale]/[screenName]/post/[id]/(engagements)/_components/PostEngagementsHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PostEngagementsHeader';
