--- src/app/[locale]/_components/HomeTabBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HomeTabBar';
